Restructure how unit modules are shipped within the package
