+++ thoth/adviser/steps/tensorflow/tf_avx2.py
