+++ thoth/adviser/wraps/pytorch/mkl_threads.txt
@@ -1,0 +1,1 @@
+# See thoth/adviser/wraps/tensorflow/mkl_threads.py
