+++ thoth/adviser/wraps/tensorflow/tf_intel.py
